+++ crates/cosmos/cosmos-wasm-relayer/src/encoding/type_url.rs
@@ -4,6 +4,7 @@ use hermes_cosmos_chain_components::types::tendermint::{
     TendermintClientState, TendermintConsensusState,
 };
 use hermes_wasm_encoding_components::components::WasmEncodingComponents;
+use hermes_wasm_encoding_components::types::client_message::WasmClientMessage;
 use hermes_wasm_encoding_components::types::client_state::WasmClientState;
 use hermes_wasm_encoding_components::types::consensus_state::WasmConsensusState;
 
@@ -19,6 +20,7 @@ delegate_components! {
         [
             WasmClientState,
             WasmConsensusState,
+            WasmClientMessage,
         ]:
             WasmEncodingComponents,