+++ crates/cosmos/cosmos-wasm-relayer/src/impls/client_state.rs
@@ -8,7 +8,7 @@ use hermes_relayer_components::chain::traits::types::height::HasHeightType;
 use ibc_relayer_types::core::ics02_client::client_state::ClientState;
 use ibc_relayer_types::Height;
 
-use crate::types::client_state::WrappedTendermintClientState;
+use crate::types::client_state::WasmTendermintClientState;
 
 pub struct ProvideWrappedTendermintClientState;
 
@@ -17,25 +17,25 @@ impl<Chain, Counterparty> ProvideClientStateType<Chain, Counterparty>
 where
     Chain: Async,
 {
-    type ClientState = WrappedTendermintClientState;
+    type ClientState = WasmTendermintClientState;
 }
 
 impl<Chain, Counterparty> ClientStateFieldsGetter<Chain, Counterparty>
     for ProvideWrappedTendermintClientState
 where
-    Chain: HasClientStateType<Counterparty, ClientState = WrappedTendermintClientState>
+    Chain: HasClientStateType<Counterparty, ClientState = WasmTendermintClientState>
         + HasHeightType<Height = Height>,
 {
-    fn client_state_latest_height(client_state: &WrappedTendermintClientState) -> Height {
+    fn client_state_latest_height(client_state: &WasmTendermintClientState) -> Height {
         client_state.tendermint_client_state.latest_height
     }
 
-    fn client_state_is_frozen(client_state: &WrappedTendermintClientState) -> bool {
+    fn client_state_is_frozen(client_state: &WasmTendermintClientState) -> bool {
         client_state.tendermint_client_state.is_frozen()
     }
 
     fn client_state_has_expired(
-        client_state: &WrappedTendermintClientState,
+        client_state: &WasmTendermintClientState,
         elapsed: Duration,
     ) -> bool {
         elapsed > client_state.tendermint_client_state.trusting_period