+++ crates/encoding/encoding-components/src/impls/with_context.rs
@@ -1,11 +1,12 @@
+use crate::traits::convert::{CanConvert, Converter};
 use crate::traits::decode::{CanDecode, Decoder};
 use crate::traits::decode_mut::{CanDecodeMut, MutDecoder};
 use crate::traits::encode::{CanEncode, Encoder};
 use crate::traits::encode_mut::{CanEncodeMut, MutEncoder};
 
-pub struct EncodeWithContext;
+pub struct WithContext;
 
-impl<Encoding, Strategy, Value> Encoder<Encoding, Strategy, Value> for EncodeWithContext
+impl<Encoding, Strategy, Value> Encoder<Encoding, Strategy, Value> for WithContext
 where
     Encoding: CanEncode<Strategy, Value>,
 {
@@ -14,7 +15,7 @@ where
     }
 }
 
-impl<Encoding, Strategy, Value> Decoder<Encoding, Strategy, Value> for EncodeWithContext
+impl<Encoding, Strategy, Value> Decoder<Encoding, Strategy, Value> for WithContext
 where
     Encoding: CanDecode<Strategy, Value>,
 {
@@ -23,7 +24,7 @@ where
     }
 }
 
-impl<Encoding, Strategy, Value> MutEncoder<Encoding, Strategy, Value> for EncodeWithContext
+impl<Encoding, Strategy, Value> MutEncoder<Encoding, Strategy, Value> for WithContext
 where
     Encoding: CanEncodeMut<Strategy, Value>,
 {
@@ -36,7 +37,7 @@ where
     }
 }
 
-impl<Encoding, Strategy, Value> MutDecoder<Encoding, Strategy, Value> for EncodeWithContext
+impl<Encoding, Strategy, Value> MutDecoder<Encoding, Strategy, Value> for WithContext
 where
     Encoding: CanDecodeMut<Strategy, Value>,
 {
@@ -45,5 +46,14 @@ where
         buffer: &mut Encoding::DecodeBuffer<'_>,
     ) -> Result<Value, Encoding::Error> {
         encoding.decode_mut(buffer)
+    }
+}
+
+impl<Encoding, From, To> Converter<Encoding, From, To> for WithContext
+where
+    Encoding: CanConvert<From, To>,
+{
+    fn convert(encoding: &Encoding, from: &From) -> Result<To, Encoding::Error> {
+        encoding.convert(from)
     }
 }