+++ crates/wasm/wasm-encoding-components/src/components.rs
@@ -5,7 +5,7 @@ pub use hermes_cosmos_encoding_components::components::{
     DecodeBufferTypeComponent, EncodeBufferTypeComponent,
 };
 use hermes_encoding_components::impls::delegate::DelegateEncoding;
-use hermes_encoding_components::impls::with_context::EncodeWithContext;
+use hermes_encoding_components::impls::with_context::WithContext;
 pub use hermes_encoding_components::traits::convert::ConverterComponent;
 pub use hermes_encoding_components::traits::decode::DecoderComponent;
 pub use hermes_encoding_components::traits::decode_mut::MutDecoderComponent;
@@ -20,11 +20,14 @@ use hermes_protobuf_encoding_components::impls::encode::buffer::EncodeProtoWithM
 use hermes_protobuf_encoding_components::impls::via_any::EncodeViaAny;
 pub use hermes_protobuf_encoding_components::traits::length::EncodedLengthGetterComponent;
 use hermes_protobuf_encoding_components::types::strategy::{ViaAny, ViaProtobuf};
+use ibc::clients::wasm_types::client_message::WASM_CLIENT_MESSAGE_TYPE_URL;
 use ibc::core::client::types::Height;
 use prost_types::Any;
 
+use crate::impls::encode::client_message::EncodeWasmClientMessage;
 use crate::impls::encode::client_state::EncodeWasmClientState;
 use crate::impls::encode::consensus_state::EncodeWasmConsensusState;
+use crate::types::client_message::WasmClientMessage;
 use crate::types::client_state::WasmClientState;
 use crate::types::consensus_state::WasmConsensusState;
 
@@ -51,7 +54,7 @@ define_components! {
         ]:
             DelegateEncoding<WasmEncodeMutComponents>,
         SchemaGetterComponent:
-            DelegateEncoding<WasmTypeUrlSchemas>,
+            WasmTypeUrlSchemas,
     }
 }
 
@@ -63,10 +66,15 @@ pub struct WasmEncoderComponents;
 
 delegate_components! {
     WasmConverterComponents {
-        (WasmClientState, Any): EncodeAsAnyProtobuf<ViaProtobuf, EncodeWithContext>,
-        (Any, WasmClientState): DecodeAsAnyProtobuf<ViaProtobuf, EncodeWithContext>,
-        (WasmConsensusState, Any): EncodeAsAnyProtobuf<ViaProtobuf, EncodeWithContext>,
-        (Any, WasmConsensusState): DecodeAsAnyProtobuf<ViaProtobuf, EncodeWithContext>,
+        [
+            (WasmClientState, Any),
+            (WasmConsensusState, Any),
+        ]: EncodeAsAnyProtobuf<ViaProtobuf, WithContext>,
+
+        [
+            (Any, WasmClientState),
+            (Any, WasmConsensusState),
+        ]: DecodeAsAnyProtobuf<ViaProtobuf, WithContext>,
     }
 }
 
@@ -81,30 +89,44 @@ delegate_components! {
 
         (ViaProtobuf, WasmConsensusState):
             EncodeWasmConsensusState,
+
+        (ViaProtobuf, WasmClientMessage):
+            EncodeWasmClientMessage,
     }
 }
 
 delegate_components! {
     WasmEncoderComponents {
-        (ViaAny, WasmClientState): EncodeViaAny<ViaProtobuf>,
-        (ViaProtobuf, WasmClientState): EncodeProtoWithMutBuffer,
-        (ViaAny, WasmConsensusState): EncodeViaAny<ViaProtobuf>,
-        (ViaProtobuf, WasmConsensusState): EncodeProtoWithMutBuffer,
+        [
+            (ViaAny, WasmClientState),
+            (ViaAny, WasmConsensusState),
+            (ViaAny, WasmClientMessage),
+        ]: EncodeViaAny<ViaProtobuf>,
+
+        [
+            (ViaProtobuf, WasmClientState),
+            (ViaProtobuf, WasmConsensusState),
+            (ViaProtobuf, WasmClientMessage),
+        ]: EncodeProtoWithMutBuffer,
     }
 }
 
 pub struct WasmTypeUrlSchemas;
 
-delegate_components! {
-    WasmTypeUrlSchemas {
-        WasmClientState: WasmClientStateUrl,
-        WasmConsensusState: WasmConsensusStateUrl,
-    }
-}
-
-impl_type_url!(WasmClientStateUrl, "/ibc.lightclients.wasm.v1.ClientState");
+impl_type_url!(
+    WasmTypeUrlSchemas,
+    WasmClientState,
+    "/ibc.lightclients.wasm.v1.ClientState",
+);
 
 impl_type_url!(
-    WasmConsensusStateUrl,
-    "/ibc.lightclients.wasm.v1.ConsensusState"
+    WasmTypeUrlSchemas,
+    WasmConsensusState,
+    "/ibc.lightclients.wasm.v1.ConsensusState",
+);
+
+impl_type_url!(
+    WasmTypeUrlSchemas,
+    WasmClientMessage,
+    WASM_CLIENT_MESSAGE_TYPE_URL,
 );