+++ crates/wasm/wasm-encoding-components/src/impls/encode/client_state.rs
@@ -2,7 +2,7 @@ use cgp::prelude::*;
 use hermes_encoding_components::impls::encode_mut::combine::CombineEncoders;
 use hermes_encoding_components::impls::encode_mut::field::EncodeField;
 use hermes_encoding_components::impls::encode_mut::from::DecodeFrom;
-use hermes_encoding_components::impls::with_context::EncodeWithContext;
+use hermes_encoding_components::impls::with_context::WithContext;
 use hermes_encoding_components::traits::decode_mut::MutDecoderComponent;
 use hermes_encoding_components::traits::encode_mut::MutEncoderComponent;
 use hermes_encoding_components::traits::transform::Transformer;
@@ -30,7 +30,7 @@ delegate_components! {
                 >,
                 EncodeField<
                     symbol!("latest_height"),
-                    EncodeLengthDelimitedProtoField<3, EncodeWithContext>,
+                    EncodeLengthDelimitedProtoField<3, WithContext>,
                 >,
             ]>,
         MutDecoderComponent: DecodeFrom<
@@ -38,7 +38,7 @@ delegate_components! {
             CombineEncoders<HList![
                 EncodeByteField<1>,
                 EncodeByteField<2>,
-                DecodeRequiredProtoField<3, EncodeWithContext>,
+                DecodeRequiredProtoField<3, WithContext>,
             ]>
         >,
     }