+++ crates/wasm/wasm-encoding-components/src/impls/mod.rs
@@ -1,2 +1,3 @@
+pub mod convert;
 pub mod encode;
 pub mod strategies;