+++ crates/wasm/wasm-encoding-components/src/types/client_message.rs
@@ -1,0 +1,3 @@
+use ibc::clients::wasm_types::client_message::ClientMessage;
+
+pub type WasmClientMessage = ClientMessage;